refactor(runtime): Move HTTP shell seams
Move the runtime-owned HTTP shell seam types into network.crypta.runtime.http.

Add endpoint-owned bridge factories for the default production bindings, update bootstrap/runtime imports, and refresh the focused tests and package docs so higher-level runtime code depends on runtime-owned seam types instead of endpoint-owned ones.
