+++ src/main/java/network/crypta/runtime/http/HttpShellRuntimeSupportFactory.java
@@ -1,4 +1,4 @@
-package network.crypta.runtime.endpoints.http;
+package network.crypta.runtime.http;
 
 import network.crypta.clients.http.HttpShellRuntimeSupport;
 import network.crypta.node.NodeClientCore;
@@ -7,10 +7,10 @@ import network.crypta.node.NodeClientCore;
  * Creates runtime-owned {@link HttpShellRuntimeSupport} instances for HTTP shell wiring.
  *
  * <p>This interface is the narrow construction seam that higher-level runtime code uses when it
- * needs HTTP shell support backed by a {@link NodeClientCore}. It lets packages outside {@code
- * network.crypta.runtime.endpoints.http} request runtime support without naming a concrete bridge
- * class directly. That keeps the endpoint package responsible for binding top-level wiring to the
- * current implementation while preserving the existing startup behavior.
+ * needs HTTP shell support backed by a {@link NodeClientCore}. It lets higher-level runtime
+ * packages request runtime support without naming a concrete bridge class directly. That keeps the
+ * composition root responsible for selecting the production binding while preserving the existing
+ * startup behavior.
  *
  * <p>Typical callers choose a factory during daemon startup, then invoke {@link
  * #create(NodeClientCore)} while assembling HTTP shell components. The interface does not define
@@ -36,20 +36,4 @@ public interface HttpShellRuntimeSupportFactory {
    *     supplied daemon core
    */
   HttpShellRuntimeSupport create(NodeClientCore core);
-
-  /**
-   * Returns the core-backed runtime support factory used by the current HTTP bridge.
-   *
-   * <p>This helper centralizes the production binding to {@link CoreHttpShellRuntimeSupport} inside
-   * the runtime HTTP endpoint package. Callers outside this package can therefore use the default
-   * runtime path without importing the concrete bridge class directly. The returned factory is
-   * suitable for repeated reuse anywhere the current node runtime should preserve the legacy
-   * core-backed behavior.
-   *
-   * @return factory that constructs {@link CoreHttpShellRuntimeSupport} instances from a supplied
-   *     daemon core
-   */
-  static HttpShellRuntimeSupportFactory coreBacked() {
-    return CoreHttpShellRuntimeSupport::new;
-  }
 }
