+++ src/main/java/network/crypta/runtime/services/NodeServicesSubsystem.java
@@ -18,8 +18,8 @@ import network.crypta.runtime.alerts.SimpleUserAlert;
 import network.crypta.runtime.alerts.TimeSkewDetectedUserAlert;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.diagnostics.DefaultNodeDiagnostics;
-import network.crypta.runtime.endpoints.http.HttpShellContainer;
-import network.crypta.runtime.endpoints.http.HttpShellContainerFactory;
+import network.crypta.runtime.http.HttpShellContainer;
+import network.crypta.runtime.http.HttpShellContainerFactory;
 import network.crypta.support.JVMVersion;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.Ticker;
