+++ src/test/java/network/crypta/runtime/bootstrap/NodeBootstrapTest.java
@@ -9,7 +9,7 @@ import network.crypta.fs.ServiceDirs;
 import network.crypta.node.Node;
 import network.crypta.node.NodeInitException;
 import network.crypta.node.ProgramDirectory;
-import network.crypta.runtime.endpoints.http.HttpShellContainer;
+import network.crypta.runtime.http.HttpShellContainer;
 import network.crypta.support.io.NativeThread;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
