+++ src/test/java/network/crypta/runtime/services/NodeServicesSubsystemTest.java
@@ -21,8 +21,8 @@ import network.crypta.runtime.alerts.TimeSkewDetectedUserAlert;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
 import network.crypta.runtime.diagnostics.DefaultNodeDiagnostics;
-import network.crypta.runtime.endpoints.http.HttpShellContainer;
-import network.crypta.runtime.endpoints.http.HttpShellContainerFactory;
+import network.crypta.runtime.http.HttpShellContainer;
+import network.crypta.runtime.http.HttpShellContainerFactory;
 import network.crypta.runtime.updater.NodeUpdateManager;
 import network.crypta.support.JVMVersion;
 import network.crypta.support.PriorityAwareExecutor;
